Add service which downloads data in .csv format
Add service which downloads zipped .csv used by Internet Stream.
Change datatype of event from Int to Long.

+++ server-kotlin/src/main/kotlin/digital/cesko/common/Tables.kt
@@ -85,7 +85,7 @@ object Payments : Table("data.payments") {
     val paragraph = integer("paragraph").nullable()
     val item = integer("item").nullable()
     val unit = integer("unit").nullable()
-    val event = integer("event").nullable()
+    val event = long("event").nullable()
     val amount = decimal("amount", 14, 2).nullable()
     val date = date("date").nullable()
     val counterpartyId = text("counterparty_id").nullable()
